% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportDrugZ.R
\name{BuildDrugZdz}
\alias{BuildDrugZdz}
\title{Build dataframe of RRA Day 0 Comparisons}
\usage{
BuildDrugZdz(drugz_list, pattern = "^([dD]|[dD]ay)0_", order = NULL)
}
\arguments{
\item{drugz_list}{List of DrugZ dataframes as imported with ImportDrugZ()}

\item{pattern}{Regex pattern to identify DrugZ day 0 comparisons (or desired
comparisons) in the drugz_list names. Default matches names starting with
D/day0_", as \code{"^([dD]|[dD]ay)0_"}.}

\item{order}{Optional character vector or factor specifying the order of
comparisons in the output dataframe.}
}
\value{
a data frame with the NormZ, minimum pvalue and FDR for each gene
across all provided DrugZ day 0 (or as indicated by the prefix) objects
}
\description{
Build dataframe of RRA Day 0 Comparisons
}
