% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportRRA.R
\name{BuildRRAdz}
\alias{BuildRRAdz}
\title{Build dataframe of RRA Day 0 Comparisons}
\usage{
BuildRRAdz(rra_list, pattern = "^[dD]|[dD]ay0_.+", order = NULL)
}
\arguments{
\item{rra_list}{List of RRA dataframes as imported with ImportRRA()}

\item{pattern}{Regex pattern to identify RRA day 0 comparisons (or desired
comparisons) in the rra_list names. Default matches names starting with "d"
or "D" or containing "day0_", as \code{"^[dD]|[dD]ay0_.+"}.}

\item{order}{Optional character vector or factor specifying the order of
comparisons in the output dataframe.}
}
\value{
a data frame with the LFC, minimum pvalue and FDR for each gene
across all provided RRA day 0 objects
}
\description{
Build dataframe of RRA Day 0 Comparisons
}
