% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crispyr-package.R
\docType{package}
\name{crispyr-package}
\alias{crispyr}
\alias{crispyr-package}
\title{crispyr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Martín González Fernández \email{martin.gonzalezfernandez@unibe.ch} (\href{https://orcid.org/0009-0006-8308-5493}{ORCID})

}
\keyword{internal}
