% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nin.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Not In operator}
\usage{
x \%nin\% y
}
\arguments{
\item{x}{Character vector to check belonging}

\item{y}{Character vector to check x against}
}
\value{
Boolean vector of the same length as x
}
\description{
Negation of the \%in\% operator for checking if elements are NOT in a vector.
}
