% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportDrugZ.R
\name{ImportDrugZ}
\alias{ImportDrugZ}
\title{Import DrugZ Results}
\usage{
ImportDrugZ(DrugZ_dir, extra_prefix = NULL)
}
\arguments{
\item{DrugZ_dir}{Path to the directory containing DrugZ results folders.}

\item{extra_prefix}{Optional prefix to add to the names of the imported dataframes.}
}
\value{
a list of imported DrugZ gene summaries with names extra_prefix\if{html}{\out{<comparison>}}
}
\description{
Import DrugZ Results
}
