% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportRRA.R
\name{ImportRRA}
\alias{ImportRRA}
\title{Import RRA Results}
\usage{
ImportRRA(RRA_dir, extra_prefix = NULL)
}
\arguments{
\item{RRA_dir}{Path to the directory containing RRA results folders.}

\item{extra_prefix}{Optional prefix to add to the names of the imported dataframes.}
}
\value{
nothing, but assigns dataframes to the global environment with names rra_\if{html}{\out{<comparison>}}
}
\description{
Import RRA Results
}
