% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_and_map.R
\name{match_constructs_approximate}
\alias{match_constructs_approximate}
\title{Approximate Barcode Matching (Internal)}
\usage{
match_constructs_approximate(
  extracted_barcodes,
  reference_library,
  max_mismatches = 1,
  quiet = FALSE
)
}
\arguments{
\item{extracted_barcodes}{Data.table. Extracted barcode information.}

\item{reference_library}{Data.table. Reference construct library.}

\item{max_mismatches}{Integer. Maximum hamming distance.}

\item{quiet}{Logical. Suppress messages.

Slower approximate matching allowing for mismatches.}
}
\description{
Approximate Barcode Matching (Internal)
}
