% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NineSquares.R
\name{NineSquares}
\alias{NineSquares}
\title{Create and annotate a Nine Squares Plot}
\usage{
NineSquares(
  data,
  control,
  treatment,
  ctrl_pval,
  treat_pval,
  min_sgrna = 3,
  min_pval,
  scale = 2,
  alpha = 0.4,
  shape = 21,
  top_labeled = 5,
  force_zero_center = c("none", "both", "control", "treatment"),
  xlab,
  ylab,
  title,
  xcut,
  ycut,
  slopecut,
  legend = FALSE,
  filename,
  groups_labeled = c("top_center", "bottom_center", "middle_right", "middle_left"),
  goi,
  goi_auto = TRUE,
  goi_shape = 21,
  goi_color = "black",
  goi_fill = "black",
  goi_size = 2.5,
  goi_label_type = c("label", "text"),
  goi_label_color = "black",
  goi_label_size = 4
)
}
\arguments{
\item{data}{input data frame containing genes, number of sgRNAs per gene,
enrichment scores and p values / p.adjusted / FDRs for all conditions you
want to compare. First column should be by default the gene id, and second
column should be number of sgRNAs.}

\item{control}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> control enrichment score
column, will be plotted on x axis. Must be numerical.}

\item{ctrl_pval}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> control pvalue /
p.adjusted / FDR column. Must be numerical.}

\item{treat_pval}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> treatment pvalue /
p.adjusted / FDR column. Must be numerical.}

\item{min_sgrna}{numeric, minimum number of sgRNAs per gene to include in the
analysis. Default is 3.}

\item{min_pval}{numeric, p-value threshold to apply to whether genes appear
in the graph or not.}

\item{scale}{numeric, standard deviation multiplier to calculate thresholds.
Default is 2.}

\item{alpha}{numeric, transparency of the points in the plot. Default is 0.4.}

\item{shape}{numeric, shape of the points in the plot. Default is 21.}

\item{top_labeled}{numeric, indicates how many top genes (by euclidean
distance) in each group/square to label. Default is 5.}

\item{force_zero_center}{character, one of "none", "both", "control", or
"treatment". Indicates if cutoffs should be forced to center around zero
instead of the mean.}

\item{xlab}{character, label for the x axis. Default is "Control Enrichment
Score".}

\item{ylab}{character, label for the y axis. Default is "Treatment Enrichment
Score".}

\item{title}{character, graph title}

\item{xcut}{2-length numerical vector indicating where cutoffs should be
established for the x axis. Replaces default calculation, which is mean ±
scale * standard deviation of control}

\item{ycut}{2-length numerical vector indicating where cutoffs should be
established for the y axis. Replaces default calculation, which is mean ±
scale * standard deviation of treatment}

\item{slopecut}{2-length numerical vector indicating where cutoffs should be
established for the x axis. Replaces default calculation, which is mean ±
scale * standard deviation of treatment - control}

\item{legend}{logical, indicate whether legend should be plotted. Default FALSE}

\item{filename}{character, path/name of the file to be written containing the data.}

\item{groups_labeled}{groups to have their top genes labeled. Defaults to "top_center", "bottom_center", "middle_right" and "middle_left"}

\item{goi}{character vector of \emph{g}enes \emph{o}f \emph{i}nterest to be highlighted.}

\item{goi_auto}{logical, indicates whether highlighted genes of interest are coloured according to the square they are in.}

\item{goi_shape}{numeric, shape of the highlighted genes of interest in the plot. Default is 21.}

\item{goi_color}{character, color aesthetic of the highlighted genes of interest in the plot. Default is black.}

\item{goi_fill}{character, fill aesthetic of the highlighted genes of interest in the plot. Default is black. goi_auto needs to be set to FALSE.}

\item{goi_size}{size, shape of the highlighted genes of interest in the plot. Default is 4.}

\item{goi_label_type}{character, one of "label" or "text". Indicates whether the annotation is text or a label. Defaults to "label".}

\item{goi_label_color}{character, color of the label text, default is "black".}

\item{goi_label_size}{character, size of the label text, default is 4.}

\item{treament}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> treatment enrichment
score column, will be plotted on y axis. Must be numerical.}
}
\value{
a ggplot2 object
}
\description{
Create and annotate a Nine Squares Plot
}
\examples{

set.seed(42)

data <- data.frame(
gene = paste0("Gene", 1:20000),
  num = sample(1:10, 20000, replace = TRUE),
  untreated_LFC = stats::rnorm(20000, mean = 0, sd = 3),
  treated_LFC = stats::rnorm(20000, mean = 0, sd = 3),
  untreated_pval = stats::runif(20000, min = 0, max = 1),
  treated_pval = stats::runif(20000, min = 0, max = 1)
)

NineSquares(data,
  untreated_LFC,
  treated_LFC,
  untreated_pval,
  treated_pval,
  min_pval = 0.05,
  scale = 2,
  shape = 15,
  top_labeled = 10,
  xlab = "Stupidly Awesome",
  ylab = "Awesomely Stupid",
  title = "No way you're doing this?",
  legend = TRUE,
  filename = "testing_whynot",
  groups_labeled = c("top_center", "bottom_center"),
  goi = c("Gene2994", "Gene405", "Gene1450", "Gene592", "Gene14261"))
}
