% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NineSquares_helpers.R
\name{NSaddgoi}
\alias{NSaddgoi}
\title{Add Genes of Interest (goi) highlights}
\usage{
NSaddgoi(
  data,
  graph,
  goi_list,
  goi_auto = FALSE,
  goi_shape = 21,
  goi_color = "black",
  goi_fill = "black",
  goi_size = 2,
  goi_label_type = c("text", "label"),
  goi_label_color = "black",
  goi_label_size = 4
)
}
\arguments{
\item{data}{processed dataframe, needs to be the same fed into NSbasegraph}

\item{graph}{graph output of basegraph}

\item{goi_list}{vector list of genes of interest}

\item{goi_auto}{logical, indicate default processing of points with slightly
darker colors than the ones in the graph. Pretttyyyyy}

\item{goi_shape}{numerical (ggplot2 aesthetics), shape of the highlighted point}

\item{goi_color}{highlighted point color, if goi_shape is between 21 and 28 it
refers to the outline.}

\item{goi_fill}{if goi_shape is between 21 and 28, inside part of the point}

\item{goi_size}{numerical, size of the highlighted point, default is 2}

\item{goi_label_type}{type of label, either "text" or "label"}

\item{goi_label_color}{color of the label text, default is black}

\item{goi_label_size}{size of the label text, default is 4}
}
\value{
highlighted ggplot graph
}
\description{
Add Genes of Interest (goi) highlights
}
