% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NineSquares_helpers.R
\name{NSaddtoplabels}
\alias{NSaddtoplabels}
\title{Add soft labels depicting the top performing (by euclidian distance) genes
for each square}
\usage{
NSaddtoplabels(
  graph,
  data,
  groups_labeled = c("top_center", "bottom_center", "middle_right", "middle_left"),
  top_labeled = 10
)
}
\arguments{
\item{graph}{graph output of basegraph}

\item{data}{processed dataframe, needs to be the same fed into NSbasegraph}

\item{groups_labeled}{square groups where top n genes will be labeled.
Default is c("top_center", "bottom_center", "middle_right", "middle_left")}

\item{top_labeled}{number of genes from the top by euclidian distance to be labeled}
}
\value{
a ggplot2 graph with labeled points
}
\description{
Add soft labels depicting the top performing (by euclidian distance) genes
for each square
}
