% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NineSquares_helpers.R
\name{NSaxislabels}
\alias{NSaxislabels}
\title{Add labels to a graph}
\usage{
NSaxislabels(graph, xlab, ylab)
}
\arguments{
\item{graph}{a graph}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}
}
\value{
a x and y labeled graph
}
\description{
Add labels to a graph
}
