% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NineSquares.R
\name{NSbasedf}
\alias{NSbasedf}
\title{Title}
\usage{
NSbasedf(data, control, treament, ctrl_pval, treat_pval, min_sgrna)
}
\arguments{
\item{data}{data frame containing genes, number of sgRNAs per gene, enrichment scores and p values / p.adjusted / FDRs
for all conditions you want to compare. First column should be by default the gene id, and second column should be number of sgRNAs}

\item{min_sgrna}{}
}
\value{
a data frame filtered by min_sgrna with selected columns
}
\description{
Title
}
