% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NineSquares_helpers.R
\name{NSbasegraph}
\alias{NSbasegraph}
\title{Generate base graph from preprocessed data frame for Nine Squares plot}
\usage{
NSbasegraph(
  data,
  x_cutoff,
  y_cutoff,
  slope_cutoff,
  size_var,
  size = 2,
  alpha = 0.4,
  shape = 21,
  legend = FALSE
)
}
\arguments{
\item{data}{data frame as generated from NSbasegraph and processed with
NSsquares, containing square information and rank per square, as well as
control and treatment enrichment value columns. Can be filtered by
min-pval.}

\item{x_cutoff}{numerical length-2 vector with x axis cutoff values, as from
NSgencutoff}

\item{y_cutoff}{numerical length-2 vector with y axis cutoff values, as from
NSgencutoff}

\item{slope_cutoff}{numerical length-2 vector with diagonal axis cutoff
values, as from NSgencutoff}

\item{size_var}{character, column name to map to point size. If missing,
constant size (specified by \code{size} parameter) is used for all points.}

\item{size}{numerical, size of geom_point, default is 2}

\item{alpha}{numerical, transparency of geom_point, default is 0.4}

\item{shape}{numerical, shape of geom_point, default is 21}

\item{legend}{logical, indicates whether you want square to be displayed in
the legend}
}
\value{
a base Nine Squares graph
}
\description{
Generate base graph from preprocessed data frame for Nine Squares plot
}
