% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NineSquares.R
\name{NScutoff}
\alias{NScutoff}
\title{Generate cutoffs for Nine Square Plots}
\usage{
NScutoff(x, scale, force_zero_center = FALSE)
}
\arguments{
\item{x}{numerical column / vector}

\item{scale}{standard deviation multiplier}

\item{force_zero_center}{indicate if you want the cutoffs to be forcefully centered around zero}
}
\value{
A vector of length 2 with the lower and upper cutoff
}
\description{
Generate cutoffs for Nine Square Plots
}
\examples{

data <- data.frame(
  gene = paste0("Gene", 1:100),
  LFC = rnorm(1000, mean = 0, sd = 2)
)

cutoffs <- NScutoff(data$LFC, scale = 1)
cutoffs
}
