% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NineSquares_postprocess.R
\name{NSextractdata}
\alias{NSextractdata}
\title{Extract data from Nine Squares plot object and assign it}
\usage{
NSextractdata(list)
}
\arguments{
\item{graph}{a Nine Squares plot object}

\item{assign}{logical, whether to assign the data to the global environment}
}
\value{
the data frame used to generate the plot
}
\description{
Extract data from Nine Squares plot object and assign it
}
