% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NineSquares_helpers.R
\name{NSfilterpval}
\alias{NSfilterpval}
\title{Filter data using minimum p value}
\usage{
NSfilterpval(data, min_pval = 0.05)
}
\arguments{
\item{data}{temporary data data frame generated with NSbasedf function}

\item{min_pval}{numerical, minimum p value to filter genes. Default is 0.05.}
}
\value{
a data frame filtered by min_pval
}
\description{
Filter data using minimum p value
}
