% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NineSquares_helpers.R
\name{NSgencutoff}
\alias{NSgencutoff}
\title{generate cutoffs using NScutoff}
\usage{
NSgencutoff(
  data,
  scale,
  force_zero_center = c("control", "treatment", "both", "none")
)
}
\arguments{
\item{data}{data frame}

\item{scale}{standard deviation multiplier.}

\item{force_zero_center}{indicate which cutoffs to be forcefully centered
around zero. Options are "control", "treatment", "both" or "none"
(default).}
}
\value{
A list with x_cutoff, y_cutoff and slope_cutoff
}
\description{
generate cutoffs using NScutoff
}
\examples{

data <- data.frame(
  gene = paste0("Gene", 1:100),
  control = stats::rnorm(1000, mean = 0, sd = 2),
  treatment = stats::rnorm(1000, mean = 0, sd = 2)
)
cutoffs <- NSgencutoff(data, scale = 1, force_zero_center = "both")
cutoffs
}
