% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NineSquares_helpers.R
\name{NSsquares}
\alias{NSsquares}
\title{Create and assign Squares in a dataset using cutoffs}
\usage{
NSsquares(data, x_cutoff, y_cutoff, slope_cutoff)
}
\arguments{
\item{data}{data frame containing treatment & control columns, as generated
by NSbasegraph}

\item{x_cutoff}{numerical length-2 vector with x axis cutoff values, as from
NSgencutoff}

\item{y_cutoff}{numerical length-2 vector with y axis cutoff values, as from
NSgencutoff}

\item{slope_cutoff}{numerical length-2 vector with diagonal axis cutoff
values, as from NSgencutoff}
}
\value{
a data frame containing three new columns, diff, square, and rank (=
highest diff per square)
}
\description{
Create and assign Squares in a dataset using cutoffs
}
