% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NineSquares_helpers.R
\name{NStitle}
\alias{NStitle}
\title{Add title to graph}
\usage{
NStitle(graph, title)
}
\arguments{
\item{graph}{a graph}

\item{title}{a title}
}
\value{
a graph with a title
}
\description{
Add title to graph
}
