% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportRRA.R
\name{BuildRRAdz}
\alias{BuildRRAdz}
\title{Build dataframe of RRA Day 0 Comparisons}
\usage{
BuildRRAdz(objects = NULL)
}
\arguments{
\item{objects}{Optional character vector of RRA day 0 object names to include.}
}
\value{
a data frame with the LFC, minimum pvalue and FDR for each gene across all provided RRA day 0 objects
}
\description{
Build dataframe of RRA Day 0 Comparisons
}
