% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crispyr-package.R
\docType{package}
\name{crispyr-package}
\alias{crispyr}
\alias{crispyr-package}
\title{crispyr: CRISPR Screen FastQ Processing and Analysis}
\description{
crispyr processes multiplexed FASTQ files from pooled CRISPR screens into construct and gene-level count matrices. It implements barcode extraction, library matching, sample demultiplexing, and quality control similar to the Broad Institute's PoolQ tool. The package outputs raw counts and log-normalized matrices suitable for downstream statistical analysis of screening hits.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/StrawHattM/crispyr}
  \item Report bugs at \url{https://github.com/StrawHattM/crispyr/issues}
}

}
\author{
\strong{Maintainer}: Martin Gonzalez Fernandez \email{martin.gonzalezfernandez@unibe.ch} (\href{https://orcid.org/0009-0006-8308-5493}{ORCID})

}
\keyword{internal}
