% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nin.R
\name{\%nin\%}
\alias{\%nin\%}
\title{/Not in/ operator}
\usage{
x \%nin\% y
}
\arguments{
\item{x}{Character vector to check belonging.}

\item{y}{Character vector to check x against.}
}
\value{
Boolean vector of the same length as x indicating if each element of x is not in y.
}
\description{
/Not in/ operator
}
\examples{
x <- c("A", "B", "C", "D", "E", "F")
y <- c("B", "A", "D", "E")

x \%nin\% y
c("FALSE", "FALSE", "TRUE", "FALSE", "FALSE", "TRUE")

}
