% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportMLE.R
\name{ImportMLE}
\alias{ImportMLE}
\title{Import MLE Results from MAGeCK-MLE}
\usage{
ImportMLE(MLE_dir = ".", prefix = "^mageck_MLE_", extra_prefix = NULL)
}
\arguments{
\item{MLE_dir}{Path to the directory containing MLE results folders.}

\item{prefix}{Prefix pattern to identify MLE result folders.}

\item{extra_prefix}{Optional prefix to add to the names of the imported dataframes.}
}
\value{
nothing, but assigns dataframes to the global environment with names mle_\if{html}{\out{<comparison>}}
}
\description{
Import MLE Results from MAGeCK-MLE
}
