% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NineSquares_helpers.R
\name{NSbasedf}
\alias{NSbasedf}
\title{Generate temporary data frame using input}
\usage{
NSbasedf(data, control, treament, ctrl_pval, treat_pval, min_sgrna = 3)
}
\arguments{
\item{data}{input data frame containing genes, number of sgRNAs per gene,
enrichment scores and p values / p.adjusted / FDRs for all conditions you
want to compare. First column should be by default the gene id, and second
column should be number of sgRNAs}

\item{control}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> control enrichment score
column, will be plotted on x axis. Must be numerical.}

\item{treament}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> treatment enrichment
score column, will be plotted on y axis. Must be numerical.}

\item{ctrl_pval}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> control pvalue /
p.adjusted / FDR column. Must be numerical.}

\item{treat_pval}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> treatment pvalue /
p.adjusted / FDR column. Must be numerical.}

\item{min_sgrna}{minimum number of sgRNAs per gene to include in the
analysis. Default is 3.}
}
\value{
a data frame filtered by min_sgrna with selected columns
}
\description{
Generate temporary data frame using input
}
\examples{
data <- data.frame(
  gene = paste0("Gene", 1:100),
  num = 4,
  untreated_LFC = stats::rnorm(100, mean = 0, sd = 2),
  treated_LFC = stats::rnorm(100, mean = 0, sd = 2),
  untreated_pval = stats::runif(100, min = 0, max = 1),
  treated_pval = stats::runif(100, min = 0, max = 1)
)

NSbasedf(
  data = data,
  control = untreated_LFC,
  treament = treated_LFC,
  ctrl_pval = untreated_pval,
  treat_pval = treated_pval,
  min_sgrna = 3
)
}
