% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Postprocess_utils.R
\name{saveNSdata}
\alias{saveNSdata}
\title{Save Nine Squares data to files}
\usage{
saveNSdata(graph_list, dir = "./graphdata/", prefix)
}
\arguments{
\item{graph_list}{A Nine Squares plot object or list of Nine Squares plot
objects}

\item{dir}{Directory path where files will be saved. Created if it doesn't
exist. Default: "./graphdata/"}

\item{prefix}{Prefix for output filenames. If not provided, will use
"NineSquaresData_<object_name>_"}
}
\description{
Save Nine Squares data to files
}
